requirements.txt
streamlit==1.32.0
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.13
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
chromadb==0.4.22
plotly==5.18.0
pandas==2.1.3
streamlit-agraph==0.0.45

